Put the bill calculator finction inside saving_estitamor and adding a testing file
