+++ Testing_file.py
@@ -1,0 +1,35 @@
+# ---------- Preparing inputs for **TESTING** -----------
+import pandas as pd
+import requests
+import os
+from saving_estimator import saving_est   # here we only import to test this function
+
+
+# ---------- Test the function saving estimator: ------------
+
+# Preparing the Load profile (kWh) - in the website this will be generated by the load_profile_estimator.py but here for
+# testing we use a profile from api.ceem.org.au
+LP = requests.get('http://api.ceem.org.au/LoadProfiles/Avg')
+LP = LP.json()
+df = pd.DataFrame.from_dict(LP, orient='columns')
+df['TS'] = pd.to_datetime(df['TS'], unit='ms')
+df = df[['TS', 'Load']]
+load_profile = df.copy()
+
+# PV Profile (kWh) - this file is the output of Jessie's functions - here we use a sample PV profile from excel file
+cwd = os.getcwd()
+xl = pd.ExcelFile(os.path.join(cwd, "PVProfile.xlsx"))
+pv_profile = xl.parse('testdata')
+pv_profile['TS'] = pd.to_datetime(pv_profile.TS)
+
+# Tariff - this will be selected by used from a list of tariffs in the dropdown list here we use a sample for testing
+Tariff_name = "Origin Flat Rate NSW (Endeavour area)"
+all_tariffs = requests.get('http://api.ceem.org.au/elec-tariffs/retail')
+all_tariffs = all_tariffs.json()
+for i in range(len(all_tariffs)):
+    if all_tariffs[i]['Name'] == Tariff_name:
+        selected_tariff = all_tariffs[i]
+
+# Testing the function for estimating saving (saving_est.py)
+
+saving_est(load_profile, pv_profile, selected_tariff)
