--- user_input_load_profile_process.py
@@ -1,1 +1,0 @@
-from bill_calculator import calc
